sorrentum_sandbox/spring2024/sorrentum_sandbox/spring2023/ml_projects/SorrIssue14_Team1_Implement_sandbox_for_Google_Trends/requirements.txt
dask
matplotlib
pandas
pmdarima
psycopg2_binary
python-dotenv
statsmodels
